Backend/requirements.txt
boto3==1.28.2
channels==3.0.4
channels-redis==4.0.0
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-jsonform==2.22.0
django-model-utils==4.5.1
django-object-actions==4.2.0
django-storages==1.14.2
Django==4.0.0
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
posthog==3.5.0
pandas==2.1.3
psycopg2==2.9.6
google-ads==21.1.0
google-analytics-admin==0.22.0
google-analytics-data==0.18.0
google-api-python-client==2.111.0
google-auth-oauthlib==1.1.0